Add osu! editor pooling support
